Found error in Coding read Description
So while doing some testing with Console.Writeline.. I found that the
while loop in each run function is never going to reach zero.

So while (proccess.Count != 0) --> it will never reach zero as we never
"Decremented the count" I'm not sure if we're suppose to or not.

I'll continue testing it.